Have excluded projects with real final costs less than $20m, rather than real initial costs less than $20m.
It was initially real initial costs less than $20m because this was the criteria for inclusion in the Investment Monitor in recent years. However, it is more desirable (wrt sample size) to define it relative to the final cost, and the bias it introduces to the sample is fine, as it counteracts the current bias towards small projects later in the pipeline (of the projects started in recent years, only small projects have been completed, as large projects take a long time to build.Given this, final cost is the better cost to filter on.